Move evaluator tests to a dedicated test file
The #554 evaluator tests merged in from #584 grew the module mirror test
file and tripped CodeScene's Low Cohesion delta gate (9.39 -> 7.11), while
uncertainty_metrics.py itself improved. Relocate TestBoundsAndErrorsEvaluators
to a small, cohesive tests/evaluate/test_uncertainty_metrics_evaluators.py so
the mirror file is no longer grown, and drop the now-unused evaluator imports.
